+++ tests/unit/commands/repos_test.py
@@ -65,14 +65,13 @@ class TestReposCommands(DustyTestCase):
         self.assertItemsEqual(get_config_value(constants.CONFIG_REPO_OVERRIDES_KEY),
                               {get_specs_repo().remote_path: self.temp_specs_path})
 
-    def test_override_then_manage_all(self):
+    @patch('dusty.commands.repos._manage_repo')
+    def test_override_then_manage_all(self, fake_manage_repo):
         override_repo('github.com/app/a', self.temp_specs_path)
-        self.assertItemsEqual(get_config_value(constants.CONFIG_REPO_OVERRIDES_KEY),
-                              {get_specs_repo().remote_path: self.temp_specs_path,
-                               'github.com/app/a': self.temp_specs_path})
+        override_repo('github.com/app/b', self.temp_specs_path)
         manage_all_repos()
-        self.assertItemsEqual(get_config_value(constants.CONFIG_REPO_OVERRIDES_KEY),
-                              {})
+        fake_manage_repo.assert_has_calls([call('github.com/app/a'), call('github.com/app/b'),
+                                           call(get_specs_repo().remote_path)], any_order=True)
 
     def test_override_repos_from_directory(self):
         override_repos_from_directory(self.temp_repos_path)